+++ goutils/usecase/standard_usecase.go
@@ -15,6 +15,7 @@ const (
 )
 
 type StandardUseCaseI interface {
+	GetDebug() bool
 	GetDb() *database.ProjectPsqlDb
 	GetLog() *zap.Logger
 	GetApiPort() string
@@ -28,16 +29,19 @@ type StandardUseCaseI interface {
 	GetAppRouter() http.Handler
 	NewApiServer() *http.Server
 	NewMetricServer() *http.Server
+	GetErrorMap() map[error]int
 }
 
 type StandardUseCaseRealization struct {
 	log         *zap.Logger
+	debug       bool
 	db          *database.ProjectPsqlDb
 	errC        chan error
 	serviceName string
 	apiPort     string
 	metricPort  string
 	allStopChan chan struct{}
+	errMap      map[error]int
 	// AppRouter value is set in each project
 	AppRouter http.Handler
 	// MetricRouter value is set in each project
@@ -46,6 +50,10 @@ type StandardUseCaseRealization struct {
 
 func (s *StandardUseCaseRealization) GetDb() *database.ProjectPsqlDb {
 	return s.db
+}
+
+func (s *StandardUseCaseRealization) GetDebug() bool {
+	return s.debug
 }
 
 func (s *StandardUseCaseRealization) GetLog() *zap.Logger {
@@ -62,6 +70,10 @@ func (s *StandardUseCaseRealization) GetServiceName() string {
 
 func (s *StandardUseCaseRealization) GetApiPort() string {
 	return s.apiPort
+}
+
+func (s *StandardUseCaseRealization) GetErrorMap() map[error]int {
+	return s.errMap
 }
 
 func (s *StandardUseCaseRealization) GetMetricPort() string {
@@ -103,14 +115,18 @@ func NewStandardUseCase(
 	serviceName string,
 	l *zap.Logger,
 	ec chan error,
+	projectErrMap map[error]int,
 ) (*StandardUseCaseRealization, error) {
 	var u = new(StandardUseCaseRealization)
 	u.log = l
 	u.serviceName = serviceName
 	u.errC = ec
+	u.debug = c.GetDebug()
 	u.apiPort = c.GetApiPort()
 	u.metricPort = c.GetMetricPort()
 	u.allStopChan = make(chan struct{}, 1)
+	u.errMap = make(map[error]int)
+	u.errMap = projectErrMap
 	return u, nil
 }