+++ changelog.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project makes best to adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+### [0.1.3] - 2026-14-09
+#### Code
+- event processor nats realization re-work
+
 ### [0.1.2] - 2026-26-08
 #### Code
 - StandardUsecase metrics bugfix
@@ -23,6 +28,7 @@ and this project makes best to adhere to [Semantic Versioning](https://semver.or
 #### Code
 - python utils
 
+[0.1.3]: https://github.com/Belobragin/bvv_utils/tree/v0.1.3
 [0.1.2]: https://github.com/Belobragin/bvv_utils/tree/v0.1.2
 [0.1.1]: https://github.com/Belobragin/bvv_utils/tree/v0.1.1
 [0.1.0]: https://github.com/Belobragin/bvv_utils/tree/v0.1.0